refactor: :recycle: inside basic route
move prometheus logic inside the basicroute module
