+++ app/basic_route.py
@@ -5,7 +5,8 @@ templates = Jinja2Templates(directory="templates")
 from . import prediction
 from . import handle_input
 import logging
-
+from prometheus_client import Counter,generate_latest, CONTENT_TYPE_LATEST
+from fastapi.responses import Response
 logger = logging.getLogger(__name__)
 
 #Pour l'instant set comme sa mais a generer en fonction du modèle en production
@@ -16,15 +17,24 @@ logger.info(f"Parametres static: {parameters}")
 parameters = handle_input.get_parameters_dynamique()
 logger.info(f"Parametres charger par le modèle : {parameters}")
 
+REQUEST_COUNT = Counter("http_requests_total", "Total number of HTTP requests")
+
 @router.get("/")
 async def prediction_page(request:Request):
+    REQUEST_COUNT.inc()
     return templates.TemplateResponse("prediction.html",{"request":request,"parametres_info":parameters})
 
 @router.post("/predict_immo")
 async def predict_immo(request:Request):
+    REQUEST_COUNT.inc()
     form = await request.form()
     data = dict(form) 
     print(data)
     prix = prediction.model_predict(data)
     print(f"prix estimer:{prix}")
     return {"price":prix[0]}
+
+
+@app.get("/metrics")
+def metrics():
+    return Response(generate_latest(), media_type=CONTENT_TYPE_LATEST)
